+++ R/describe.R
@@ -242,7 +242,7 @@ describe_cat <- function(data, var, max_cat = 10, out = "text", margin = 0) {
 #'
 #' @param data A dataset
 #' @param out Output format ("small"|"large")
-#' @return Dataset
+#' @return Dataset (tibble)
 #' @import dplyr
 #' @examples
 #' describe_all(iris)
@@ -271,7 +271,7 @@ describe_all <- function(data = NA, out = "large") {
   max <- NULL
 
   # define result data.frame
-  result <- data.frame(variable = character(),
+  result <- tibble::tibble(variable = character(),
                        type = character(),
                        na = integer(),
                        na_pct = double(),
@@ -322,7 +322,7 @@ describe_all <- function(data = NA, out = "large") {
     } # if
 
     result <- rbind(result,
-                    data.frame(variable = var_name,
+                    tibble::tibble(variable = var_name,
                                type = var_type,
                                na = var_na,
                                na_pct = var_na_pct,